refactor(file-system-plugin): extract shared file transfer helpers
Moves file import/export path validation, overwrite preflight, and adapter
read/write calls into reusable file transfer helpers.

Keeps the DevTools runtime handlers focused on capability checks, provider
resolution, and result event wiring while preserving the existing panel
transfer behavior.
